registration, authorization and collections for special users

+++ backend/routes/collections.js
@@ -15,32 +15,37 @@ const connection = mysql.createConnection({
 connection.connect();
 
 router.get('/collections', utils.authRequired, function (req, res, next) {
-    connection.query('SELECT * FROM collections', function (err, result) {
-        res.send({...result})
+    connection.query("SELECT userID FROM sessions WHERE sessionID = ?", [req.header("Authorization")], (e, r) => {
+        let uid = r[0]['userID']
+        connection.query('SELECT * FROM collections WHERE user = ?', [uid], (err, result) => {
+            res.send({...result})
 
-        console.log(err);
-        //console.log(result); // собственно данные
-    });
+            console.log(err);
+        });
+    })
+
 })
 
+router.post('/collections', utils.authRequired, function (req, res, next) {
 
-
-router.post('/collections', function (req, res, next) {
-
-    let data = [req.body.id, req.body.name, req.body.description];
-
-    connection.query('INSERT INTO collections (id, name, description) VALUES(?, ?, ?) ', data, function (err, results, fields) {
-        !err ? res.json(results) : res.json(err);
+    connection.query("SELECT userID FROM sessions WHERE sessionID = ?", [req.header("Authorization")], (e, r) => {
+        let uid = r[0]['userID']
+        let data = [req.body.name, req.body.description, uid];
+        connection.query('INSERT INTO collections (name, description, user) VALUES(?, ?, ?) ', data, function (err, results, fields) {
+            !err ? res.json(results) : res.json(err);
+        })
     })
 })
 
 
-router.post('/deleteCollection', function (req, res, next) {
+router.post('/deleteCollection', utils.authRequired, function (req, res, next) {
 
     let collectionId = Object.keys(req.body);
-
-    connection.query('DELETE FROM collections WHERE id IN (?) ', collectionId, function (err, results, fields) {
-        !err ? res.json(results) : res.json(err);
+    connection.query("SELECT userID FROM sessions WHERE sessionID = ?", [req.header("Authorization")], (e, r) => {
+        let uid = r[0]['userID']
+        connection.query('DELETE FROM collections WHERE id IN (?) and user = ? ', [collectionId, uid], function (err, results, fields) {
+            !err ? res.json(results) : res.json(err);
+        })
     })
 
 })
